+++ client/client.rs
@@ -187,8 +187,12 @@ impl Client {
                 Ok(false)
             }
 
-            SubCommand::Edit { task_id, path } => {
-                let message = edit(&mut self.stream, *task_id, *path).await?;
+            SubCommand::Edit {
+                task_id,
+                path,
+                label,
+            } => {
+                let message = edit(&mut self.stream, *task_id, *path, *label).await?;
                 self.handle_response(message);
                 Ok(true)
             }